[SPOT-293][REFACTOR] 투두, 알림 기존 코드 legacy 패키지 하위로 이동
