+++ src/test/java/com/example/spot/service/notification/NotificationQueryServiceTest.java
@@ -13,9 +13,9 @@ import com.example.spot.notification.domain.Notification;
 import com.example.spot.notification.domain.enums.NotifyType;
 import com.example.spot.study.domain.Study;
 import com.example.spot.notification.domain.NotificationRepository;
-import com.example.spot.notification.application.notification.NotificationQueryServiceImpl;
-import com.example.spot.notification.presentation.dto.notification.NotificationResponseDTO.NotificationListDTO;
-import com.example.spot.notification.presentation.dto.notification.NotificationResponseDTO.StudyNotificationListDTO;
+import com.example.spot.notification.application.legacy.NotificationQueryServiceImpl;
+import com.example.spot.notification.presentation.legacy.dto.notification.NotificationResponseDTO.NotificationListDTO;
+import com.example.spot.notification.presentation.legacy.dto.notification.NotificationResponseDTO.StudyNotificationListDTO;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
