+++ src/test/java/com/example/spot/service/todo/ToDoCommandServiceTest.java
@@ -10,16 +10,16 @@ import static org.mockito.Mockito.when;
 import com.example.spot.common.api.exception.handler.StudyHandler;
 import com.example.spot.member.domain.Member;
 import com.example.spot.study.domain.association.StudyMember;
-import com.example.spot.todo.application.ToDoCommandServiceImpl;
+import com.example.spot.todo.application.legacy.ToDoCommandServiceImpl;
 import com.example.spot.todo.domain.ToDo;
 import com.example.spot.study.domain.Study;
 import com.example.spot.member.domain.MemberRepository;
 import com.example.spot.study.domain.repository.StudyMemberRepository;
 import com.example.spot.study.domain.StudyRepository;
 import com.example.spot.todo.domain.ToDoRepository;
-import com.example.spot.todo.presentation.dto.request.ToDoListRequestDTO.ToDoListCreateDTO;
-import com.example.spot.todo.presentation.dto.response.ToDoListResponseDTO.ToDoListCreateResponseDTO;
-import com.example.spot.todo.presentation.dto.response.ToDoListResponseDTO.ToDoListUpdateResponseDTO;
+import com.example.spot.todo.presentation.legacy.dto.request.ToDoListRequestDTO.ToDoListCreateDTO;
+import com.example.spot.todo.presentation.legacy.dto.response.ToDoListResponseDTO.ToDoListCreateResponseDTO;
+import com.example.spot.todo.presentation.legacy.dto.response.ToDoListResponseDTO.ToDoListUpdateResponseDTO;
 import java.time.LocalDate;
 import java.util.Collections;
 import java.util.Optional;
