+++ src/test/java/com/example/spot/service/todo/ToDoQueryServiceTest.java
@@ -3,13 +3,13 @@ package com.example.spot.service.todo;
 import com.example.spot.common.api.exception.GeneralException;
 import com.example.spot.member.domain.Member;
 import com.example.spot.study.domain.association.StudyMember;
-import com.example.spot.todo.application.ToDoQueryServiceImpl;
+import com.example.spot.todo.application.legacy.ToDoQueryServiceImpl;
 import com.example.spot.todo.domain.ToDo;
 import com.example.spot.study.domain.enums.StudyApplicationStatus;
 import com.example.spot.study.domain.Study;
 import com.example.spot.study.domain.repository.StudyMemberRepository;
 import com.example.spot.todo.domain.ToDoRepository;
-import com.example.spot.todo.presentation.dto.response.ToDoListResponseDTO.ToDoListSearchResponseDTO;
+import com.example.spot.todo.presentation.legacy.dto.response.ToDoListResponseDTO.ToDoListSearchResponseDTO;
 import java.time.LocalDate;
 import java.util.Collections;
 import java.util.List;
